CM-1633: Moved FileAsset into assets package for better clarity.
