+++ comet-java-client/src/main/java/ml/comet/experiment/asset/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Contains all interfaces describing the Comet assets.
+ * Contains all interfaces and classes describing the Comet assets.
  */
 package ml.comet.experiment.asset;
